+++ src/hash_table.c
@@ -3,11 +3,11 @@
 
 struct hash_table_t;
 // struct hash_table_t {
-//     unsigned HASHSIZE;
+//     uint64_t HASHSIZE;
 //     struct linked_list_node_t** hashtab;
 // };
 
-struct hash_table_t* new_hash_table(unsigned HASHSIZE) {
+struct hash_table_t* new_hash_table(uint64_t HASHSIZE) {
     if(HASHSIZE == 0) {
         perror("Can't create a hash table with HASHSIZE=0");
         exit(1);
@@ -34,7 +34,7 @@ void free_hash_table(struct hash_table_t* a) {
     free(a);
 }
 
-struct linked_list_node_t* hash_table_look_up(struct hash_table_t* a, void* obj, unsigned (*hash)(void*), int (*cmp)(void*, void*)) {
+struct linked_list_node_t* hash_table_look_up(struct hash_table_t* a, void* obj, uint64_t (*hash)(void*), int (*cmp)(void*, void*)) {
     int slot = hash(obj) % (a->HASHSIZE);
     struct linked_list_node_t* ptr;
     for(ptr = a->hashtab[slot]; ptr; ptr = ptr->next) {
@@ -45,7 +45,7 @@ struct linked_list_node_t* hash_table_look_up(struct hash_table_t* a, void* obj,
     return NULL;
 }
 
-struct linked_list_node_t* hash_table_put(struct hash_table_t* a, void* obj, unsigned (*hash)(void*), int (*cmp)(void*, void*)) {
+struct linked_list_node_t* hash_table_put(struct hash_table_t* a, void* obj, uint64_t (*hash)(void*), int (*cmp)(void*, void*)) {
     struct linked_list_node_t* ptr = hash_table_look_up(a, obj, hash, cmp);
     int slot;
     if(ptr == NULL) {
